refactor: Convert into functional component
