+++ src/components/FooterLogo/FooterLogo.spec.jsx
